Import of XPIX/Tk-DBI-Tree-0.04 from CPAN.
gitpan-cpan-distribution: Tk-DBI-Tree
gitpan-cpan-version:      0.04
gitpan-cpan-path:         XPIX/Tk-DBI-Tree-0.04.tar.gz
gitpan-cpan-author:       XPIX
gitpan-cpan-maturity:     released

+++ demos/test.pl
@@ -21,8 +21,10 @@ my $dbh = DBI->connect(
 		or die ("Can't connect to database:", $! );
 
 END{
-	$dbh->do('DROP TABLE food');
-	$dbh->disconnect; 
+	if(defined $dbh) {
+		$dbh->do('DROP TABLE food');
+		$dbh->disconnect; 
+	}
 }
 
 while(<DATA>)
